Allow override on where the user is authenticated.
By default if no auth_source is passed, it will use the database the user is connecting too.

For some cloud providers (like Mongo Atlas), the user must be authenticated against 'admin' database even though they are accessing some other db.

The goal is to have a backwards compatible setting such that if the current users don't have this setting, it will continue to work.

Open to ideas if there is a better way to achieve this :)